+++ config.c
@@ -53,6 +53,9 @@ conf_to_color(const struct yml_node *node)
     uint16_t blue = hex_byte(&hex[4]);
     uint16_t alpha = hex_byte(&hex[6]);
 
+    if (alpha == 0)
+        return (pixman_color_t){0, 0, 0, 0};
+
     alpha |= alpha << 8;
     int alpha_div = 0xffff / alpha;